Probé la aplicacion con los cambios y todo me funciona igualmente.

+++ Proyecto_LDPII/src/main/java/com/cibertec/app/repository/EstablecimientoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.cibertec.app.entity.Establecimiento;
 
 @Repository
-public interface EstablecimientoRepository extends JpaRepository<Establecimiento, Integer> {
+public interface EstablecimientoRepository extends JpaRepository<Establecimiento, Long> {
 
 }
