+++ Proyecto_LDPII/src/main/java/com/cibertec/app/repository/FacturaRepository.java
@@ -8,9 +8,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface FacturaRepository extends JpaRepository<Factura, Integer> {
+public interface FacturaRepository extends JpaRepository<Factura, Long> {
 	@Query(value = "Select f.idFactura, f.descuento, f.precioFinal, f.fecha, f.idEvento from factura as f JOIN evento as e ON e.idEvento = f.idEvento JOIN cliente as c ON e.idCliente = c.idCliente\r\n"
 			+ "	WHERE e.idCliente = :idCliente", nativeQuery = true)
-	public List<Factura> obtenerFacturaByCliente(@Param("idCliente") Integer idCliente);
+	public List<Factura> obtenerFacturaByCliente(@Param("idCliente") Long idCliente);
 }
